requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
beautifulsoup4
ipykernel
lxml
yt-dlp
tqdm
sentence-transformers
umap-learn